Jobs minigame
Core implemented, all that needs to be done now is the rest of the NPCs table base and thats all.

Also will most likely merge all gathering tasks into one class, as they are the same for all NPCs and will have a seperate slaying one, to be done later though.