Player script fixes, and bankAll scriptAPI function added
Fixed the bank window exploit in the coal miner, Draynor willows, lobster catcher, Seers magic, shark catcher, and Varrock essence scripts.

Made it so that the coal miner script banks its entire inventory instead of just the coal ore, this way it should no longer get stuck if its inventory is full of mithril ore and uncut gems.

Fixed the Draynor willows and Seers magic trees scripts so that the bot actually chops down the trees instead of just standing next to them.

Added a bankAll function to the scriptAPI, the coal miner script is dependent on this change