Fix hang when depositing inventory or equipped items
This commit fixes an issue where the DumpContainer dialogue hangs when
a player attempts to deposit their inventory or equipped items when they
do not have any items in their inventory or are not wearing any items.

This commit also converts the DumpContainer dialogue plugin from Java to
Kotlin to work towards converting the 2009scape codebase to Kotlin.