make fight cave npcs count as tzhaar task
right now fight caves nps do not count towards txhaar tasks. Adding them to Tasks.TZHAAR is not enough since NPC.java has a check for `getZoneMonitor().handleDeath(killer)` before checking slater task.
I swaped the order of the zone check and slater taks check, which means any zone that has a specific zone handler will now grant slater xp on death. WildernessZone already had the exact same slayer code in it, which would now be excecuted in NPC.javam so it has to be taken out in order for kills not to count as double.